float版本，姿态环稳定，还有极限角度抖动问题待解决

+++ HARDWARE/app.c
@@ -291,8 +291,8 @@ void ctrl_angular_velocity(float target_angular_velocity_x,float target_angular_
 	//机体误差映射到三个电机上
 	float sin_roll, cos_roll;
 	float sin_pitch, cos_pitch;
-	sin_roll = sin(roll_encoder);cos_roll = cos(roll_encoder);
-	sin_pitch = sin(pitch_encoder);cos_pitch = cos(pitch_encoder);
+	sin_roll = sinf(roll_encoder);cos_roll = cosf(roll_encoder);
+	sin_pitch = sinf(pitch_encoder);cos_pitch = cosf(pitch_encoder);
 	float inv_cos_roll = 1.0f / cos_roll;
 	
 	float error_motor_speed_roll = cos_pitch * error_angular_velocity_body_x + sin_pitch * error_angular_velocity_body_z;
@@ -313,9 +313,9 @@ void ctrl_Attitude(void)
 	//获取当前四元数的Pitch Roll分量四元数
 	float current_quat_PR[4];
 //	float AirframeQuat[4] = {q[0],q[1],q[2],q[3]};
-	float Yaw = atan2( 2.0f*(q[0]*q[3]+q[1]*q[2]) , 1.0f-2.0f*(q[2]*q[2]+q[3]*q[3]) );
+	float Yaw = atan2f( 2.0f*(q[0]*q[3]+q[1]*q[2]) , 1.0f-2.0f*(q[2]*q[2]+q[3]*q[3]) );
 	float half_sinYaw, half_cosYaw;
-	half_sinYaw = sin(0.5f*Yaw); half_cosYaw = cos(0.5f*Yaw);
+	half_sinYaw = sinf(0.5f*Yaw); half_cosYaw = cosf(0.5f*Yaw);
 //	float YawQuat[4] = {half_cosYaw,0,0,-half_sinYaw};
 	
 	current_quat_PR[0] = q[0]*half_cosYaw + q[3]*half_sinYaw;
@@ -323,7 +323,7 @@ void ctrl_Attitude(void)
 	current_quat_PR[2] = q[2]*half_cosYaw - q[1]*half_sinYaw;
 	current_quat_PR[3] = q[3]*half_cosYaw - q[0]*half_sinYaw;
 	
-	float q_norm = sqrt(current_quat_PR[0]*current_quat_PR[0] + current_quat_PR[1]*current_quat_PR[1] + current_quat_PR[2]*current_quat_PR[2] + current_quat_PR[3]*current_quat_PR[3]);
+	float q_norm = sqrtf(current_quat_PR[0]*current_quat_PR[0] + current_quat_PR[1]*current_quat_PR[1] + current_quat_PR[2]*current_quat_PR[2] + current_quat_PR[3]*current_quat_PR[3]);
 	current_quat_PR[0] = current_quat_PR[0] / q_norm;
 	current_quat_PR[1] = current_quat_PR[1] / q_norm;
 	current_quat_PR[2] = current_quat_PR[2] / q_norm;
@@ -350,16 +350,16 @@ void ctrl_Attitude(void)
 	//使用目标角度构造目标四元数
 	float target_quat_PR[4];
 	float half_sinR, half_cosR;
-	half_sinR = sin(0.5f*target_Roll); half_cosR = cos(0.5f*target_Roll);
+	half_sinR = sinf(0.5f*target_Roll); half_cosR = cosf(0.5f*target_Roll);
 	float half_sinP, half_cosP;
-	half_sinP = sin(0.5f*target_Pitch); half_cosP = cos(0.5f*target_Pitch);
+	half_sinP = sinf(0.5f*target_Pitch); half_cosP = cosf(0.5f*target_Pitch);
 	
 	target_quat_PR[0] = half_cosR*half_cosP;
 	target_quat_PR[1] = half_cosP*half_sinR;
 	target_quat_PR[2] = half_cosR*half_sinP;
 	target_quat_PR[3] = -half_sinR*half_sinP;
 	
-	q_norm = sqrt(target_quat_PR[0]*target_quat_PR[0] + target_quat_PR[1]*target_quat_PR[1] + target_quat_PR[2]*target_quat_PR[2] + target_quat_PR[3]*target_quat_PR[3]);
+	q_norm = sqrtf(target_quat_PR[0]*target_quat_PR[0] + target_quat_PR[1]*target_quat_PR[1] + target_quat_PR[2]*target_quat_PR[2] + target_quat_PR[3]*target_quat_PR[3]);
 	target_quat_PR[0] = target_quat_PR[0] / q_norm;
 	target_quat_PR[1] = target_quat_PR[1] / q_norm;
 	target_quat_PR[2] = target_quat_PR[2] / q_norm;
@@ -380,7 +380,7 @@ void ctrl_Attitude(void)
 	q_error[2] = current_quat_PR[0]*target_quat_PR[2] - current_quat_PR[2]*target_quat_PR[0] - current_quat_PR[1]*target_quat_PR[3] + current_quat_PR[3]*target_quat_PR[1];
 	q_error[3] = current_quat_PR[0]*target_quat_PR[3] + current_quat_PR[1]*target_quat_PR[2] - current_quat_PR[2]*target_quat_PR[1] - current_quat_PR[3]*target_quat_PR[0];
 	
-	q_norm = sqrt(q_error[0]*q_error[0] + q_error[1]*q_error[1] + q_error[2]*q_error[2] + q_error[3]*q_error[3]);
+	q_norm = sqrtf(q_error[0]*q_error[0] + q_error[1]*q_error[1] + q_error[2]*q_error[2] + q_error[3]*q_error[3]);
 	q_error[0] = q_error[0] / q_norm;
 	q_error[1] = q_error[1] / q_norm;
 	q_error[2] = q_error[2] / q_norm;
@@ -388,14 +388,14 @@ void ctrl_Attitude(void)
 	
 	//计算误差旋转向量
 	float PR_rotation[3];
-	float theta = 2.0f* acos( q_error[0] );
+	float theta = 2.0f* acosf( q_error[0] );
 	if(theta > PI)
 		theta -= 2.0f*PI;
-	float sin_half_theta = sqrt( 1.0f - q_error[0]*q_error[0] );
+	float sin_half_theta = sqrtf( 1.0f - q_error[0]*q_error[0] );
 //	float scale = theta / sin_half_theta;
 	float scale;
 	if (fabsf(sin_half_theta) < FLT_EPSILON)
-		scale = 0.5;
+		scale = 0.5f;
 	else
 		scale = theta / sin_half_theta;
 	PR_rotation[0] = q_error[1] * scale * RAD2DEG;
